Task 4/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.19.0
aiohttp==3.9.1
pandas==2.1.4
